Add Descriptions to Glycan Compositions

This function adds the following columns to the variable information tibble:

'n_hex': number of Hex

'n_hexnac': number of HexNAc

'n_fuc': number of Fuc

'n_neuac': number of NeuAc

'n_neugc': number of NeuGc

'n_sia': number of Sia

Usage

add_comp_descriptions(exp)

Arguments

exp

An glyexp::experiment() object.

Value

The experiment object with the new columns added.

Details

n_sia is the sum of n_neuac and n_neugc.

The function relies on the glycan_composition column in the var_info tibble, and it should be a glyrepr_composition object. If you use the glyread package to directly read results from search engines (such as pGlyco3), this column will be automatically added to var_info. If you construct the glyexp_experiment object yourself, you need to ensure that this column exists and its type is glyrepr_composition. See glyrepr::glycan_composition() or glyrepr::as_glycan_composition() for how to create such objects manually.