Add Glycan Descriptions

This function adds glycan description columns to the variable information tibble of an glyexp::experiment() object. If structure information is available, both composition and structure descriptions are added. Otherwise, only composition descriptions are added.

Usage

add_glycan_descriptions(exp)

Arguments

exp

An glyexp::experiment() object.

Value

The experiment object with the new columns added.

Details

This function is a wrapper around add_comp_descriptions() and add_struct_descriptions().

See also