Count How Many Times a Glycan has the Given Motif

This function is closely related to has_motif(). However, instead of returning a logical value, it returns the number of times the glycan has the motif.

Usage

counts_motif(glycan, motif, alignment = NULL, ignore_linkages = FALSE)

Arguments

glycan

A 'glycan_graph' object, or an IUPAC-condensed structure string.

motif

A 'glycan_graph' object, an IUPAC-condensed structure string, or a known motif name (use available_motifs() to see all available motifs).

alignment

A character string. Possible values are "substructure", "core", "terminal" and "whole". If not provided, the value will be decided based on the motif argument. If motif is a motif name, the alignment in the database will be used. Otherwise, "substructure" will be used.

ignore_linkages

A logical value. If TRUE, linkages will be ignored in the comparison.

Value

An integer indicating how many times the glycan has the motif.

Details

This function actually perform v2f algorithm to get all possible matches between glycan and motif. However, the result is not necessarily the number of matches.

Think about the following example:

glycan: Gal(b1-?)[Gal(b1-?)]GlcNAc(b1-4)GlcNAc

motif: Gal(b1-?)[Gal(b1-?)]GlcNAc(b1-)

To draw the glycan out:

Gal 1
   \ b1-? b1-4
    GlcNAc -- GlcNAc
   / b1-?
Gal 2

To draw the motif out:

Gal 1
   \ b1-?
    GlcNAc b1-
   / b1-?
Gal 2

To differentiate the galactoses, we number them as "Gal 1" and "Gal 2" in both the glycan and the motif. The v2f subisomorphic algorithm will return two matches:

Gal 1 in the glycan matches Gal 1 in the motif, and Gal 2 matches Gal 2.

Gal 1 in the glycan matches Gal 2 in the motif, and Gal 2 matches Gal 1.

However, from a biological perspective, the two matches are the same. This function will take care of this, and return the "unique" number of matches.

For other details about the handling of monosaccharide, linkages, alignment, substituents, and implementation, see has_motif().

See also

Examples

counts_motif("Gal(b1-3)Gal(b1-3)GalNAc", "Gal(b1-")
#> [1] 2
counts_motif(
  "Man(b1-?)[Man(b1-?)]GalNAc(b1-4)GlcNAc",
  "Man(b1-?)[Man(b1-?)]GalNAc"
)
#> [1] 1
counts_motif("Gal(b1-3)Gal", "Man")
#> [1] 0