Describe N-Glycans Properties

Extract key properties of N-glycans, including:

"glycan_type": N-glycan type: high mannose, hybrid, complex, or paucimannose.

"bisecting": Bisecting GlcNAc presence.

"antennae": Number of antennae.

"core_fuc": Number of core fucoses.

"arm_fuc": Number of arm fucoses.

"terminal_gal": Number of terminal galactoses.

Usage

describe_n_glycans(glycans, strict = FALSE, parallel = FALSE)

Arguments

glycans

A list of glycan_graph objects, or a character vector of IUPAC-condensed structure strings.

strict

A logical value. If TRUE, the glycan must have "concrete" monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information. If FALSE, the function is more lenient, checking monosaacharide identities on the "simple" level (e.g. "H", "N", "F") and ignoring linkage information. Default is FALSE. This is preferred because in most cases the structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.

parallel

A logical value. If TRUE, the function will use parallel processing. Remember to call future::plan() before using this argument, otherwise the function will still use sequential processing.

Value

A tibble with the following columns: "glycan_type", "bisecting", "antennae", "core_fuc", "arm_fuc", "terminal_gal". If the input glycans have names, the tibble will have a "glycan" column. Otherwise, if IUPAC condensed strings are used, they will be used as the "glycan" column.

Details

This function is designed to work with N-glycans only. If the glycans are not N-glycans, an error is thrown.

Strictness

By default (strict = FALSE), the function is very lenient for motif checking. It only checks the monosaccharide types on the "simple" level (e.g. "H", "N", "F"), and ignores linkage information. This is preferred because in most cases the structural resolution could not be high, e.g. in most glycoproteomics studies. However, the glycans are guaranteed to be N-glycans by the glycosylation sites. In this case, we could make some assumptions about the glycan structures, and extract the key properties. For example, an H-N terminal motif is considered a terminal galactose. If you have high-resolution glycan structures, you can set strict = TRUE.

Enabling parallel processing

This function can spend a lot of time on large datasets (e.g. > 500 glycans). To speed up, you can enable parallel processing by setting parallel = TRUE. However, changing the argument to TRUE only set the function "ready" for parallel processing. You still need to call future::plan() to change the parallel backend. For example, to use the "multisession" backend:

library(future)
old_plan <- future::plan("multisession")  # Save the old plan
describe_n_glycans(glycans, parallel = TRUE)
future::plan(old_plan)  # Restore the old plan

Examples

library(glyparse)
library(purrr)

glycans <- c(
  "(N(F)(N(H(H(N))(H(N(H))))))",
  "(N(N(H(H)(H(H)(H)))))",
  "(N(F)(N(H(H(N))(H(N(H(H)))))))",
  "(N(N(H(N)(H(N(H)(F)))(H(N(H)(F))(N(H)(F))))))",
  "(N(N(H(H(N(H(A))))(H(N(H(A)))))))"
)
glycans <- map(glycans, parse_pglyco_struc)
describe_n_glycans(glycans)
#> # A tibble: 5 × 6
#>   glycan_type bisecting antennae core_fuc arm_fuc terminal_gal
#>   <chr>       <lgl>        <int>    <int>   <int>        <int>
#> 1 complex     FALSE            2        1       0            1
#> 2 highmannose FALSE           NA        0       0            0
#> 3 complex     FALSE            2        1       0            0
#> 4 complex     TRUE             3        0       3            3
#> 5 complex     FALSE            2        0       0            0