Check if a Glycan is an N-Glycan

This function checks if a glycan has the N-glycan core motif:

Man
   \
    Man - GlcNAc - GlcNAc
   /
Man

Usage

is_n_glycan(glycan, strict = FALSE)

Arguments

glycan

A glycan_graph object, or a character string of IUPAC condensed format.

strict

A logical value. If TRUE, the glycan must have "concrete" monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information. If FALSE, the function is more lenient, checking monosaacharide identities on the "generic" level (e.g. "Hex", "HexNAc") and ignoring linkage information. Default is FALSE. This is preferred because in most cases the structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.

Value

A logical value.