Number of Antennae

The number of antennae is the number of branching GlcNAc to the core mannoses in a complex N-glycan. Bisecting GlcNAc is not counted as an antenna. This functions returns NA_integer_ for non-complex N-glycans.

Usage

n_antennae(glycan, strict = FALSE)

Arguments

glycan

A glycan_graph object, or a character string of IUPAC condensed format.

strict

A logical value. If TRUE, the glycan must have "concrete" monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information. If FALSE, the function is more lenient, checking monosaacharide identities on the "simple" level (e.g. "H", "N", "F") and ignoring linkage information. Default is FALSE. This is preferred because in most cases the structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.

Value

An integer of the number of antennae.