Check if the Glycan(s) have the Given Motif(s)

These are vectorized versions of has_motif().

has_motif() checks if a motif is present in a glycan.

has_motifs() checks if a list of motifs are present in a glycans.

have_motif() checks if a motif is present in a list of glycans.

have_motifs() checks if a list of motifs are present in a list of glycans.

They can be told apart by English grammar:

Functions starting with "has" work with one glycan, "have" with multiple glycans.

Functions ending with "motif" work with one motif, "motifs" with multiple motifs.

Usage

has_motifs(
  glycan,
  motifs = NULL,
  alignments = "substructure",
  ignore_linkages = FALSE
)

have_motif(glycans, motif, alignment = "substructure", ignore_linkages = FALSE)

have_motifs(
  glycans,
  motifs = NULL,
  alignments = "substructure",
  ignore_linkages = FALSE,
  simplify = FALSE
)

Arguments

glycan

A 'glycan_graph' object, or an IUPAC-condensed structure string.

motifs

A list of 'glycan_graph' objects, a character vector of IUPAC-condensed structure strings, or a character vector of known motif names. If missing, all known motifs in the GlycoMotif GlyGen Collection will be used.

alignments

A character vector of alignment names for the motifs, either of the same length as motifs, or a single character string. If not provided and motifs are known motif names or missing (using all known motifs), the alignments in the GlycoMotif GlyGen Collection will be used.

ignore_linkages

A logical value. If TRUE, linkages will be ignored in the comparison.

glycans

A list of 'glycan_graph' objects, or a character vector of IUPAC-condensed structure strings.

motif

A 'glycan_graph' object, an IUPAC-condensed structure string, or a known motif name.

alignment

A single character, one of 'substructure', 'core', 'terminal' or 'whole'. If not provided and motif is a known motif name, the alignment in the GlycoMotif GlyGen Collection will be used.

simplify

A logical value. Only used in have_motifs(). If TRUE, drop columns (motifs) with all FALSE values. Default is FALSE.

Value

#' has_motifs() and have_motif() return a logical vector of the same length as the input motifs or glycans, respectively. have_motifs() returns a matrix of logical values, indicating if each glycan has each motif. Rows are glycans and columns are motifs.

The dimension names are determined by the input arguments. If glycans or motifs have names, they will be preserved in the result. If not, but glycans or motifs are character vectors, they will be used as names.

Details

When using known motifs in the GlycoMotif GlyGen Collection, the best practice is to not provide the alignment arguments, and let the function decide the alignment based on the motif name. However, it is still possible to override the default alignments. In this case, the user-provided alignments will be used, but a warning will be issued.

Internally, all glycan and motif inputs are converted to 'glycan_graph' objects. Therefore, if performance matters and you need to repeatedly use the same glycans or motifs, it is recommended to convert them to 'glycan_graph' objects beforehand. get_motif_graph() can be used to get the 'glycan_graph' object of a known motif. glyparse::parse_iupac_condensed() can be used to convert IUPAC-condensed structure strings to 'glycan_graph' objects.

See also

Examples

library(glyparse)

# Check if a glycan has the given motifs using names.
# Note that the `motifs` vector will become names of the result vector.
glycan <- "Gal(b1-3)GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-"
has_motifs(glycan, c("O-Glycan core 1", "O-Glycan core 2"))
#> O-Glycan core 1 O-Glycan core 2 
#>            TRUE           FALSE 

# Using IUPAC-condensed structure strings.
# Note that names of `motifs` are preserved in the result vector.
motifs <- c(M1 = "Gal(b1-3)GalNAc(a1-", M2 = "GlcNAc(b1-4)GlcNAc(a1-")
has_motifs(glycan, motifs)
#>    M1    M2 
#>  TRUE FALSE 

# Or using 'glycan_graph' objects.
motifs <- list(
  M1 = parse_iupac_condensed("Gal(b1-3)GalNAc(a1-"),
  M2 = parse_iupac_condensed("GlcNAc(b1-4)GlcNAc(a1-")
)
has_motifs(glycan, motifs)
#>    M1    M2 
#>  TRUE FALSE 

# If motifs are missing, use all known motifs in the
# GlycoMotif GlyGen Collection.
has_motifs(glycan)[1:5]
#> Blood group H (type 2) - Lewis y                        i antigen 
#>                            FALSE                            FALSE 
#>                         LacdiNAc                              GT2 
#>                            FALSE                            FALSE 
#> Blood group B (type 1) - Lewis b 
#>                            FALSE 

# Check if a list of glycans have the given motif.
glycans <- c(G1 = "Gal(b1-3)GalNAc(a1-", G2 = "GlcNAc(b1-4)GlcNAc(a1-")
have_motif(glycans, "O-Glycan core 1")
#>    G1    G2 
#>  TRUE FALSE 

# Check if a list of glycans have the given motifs.
glycans <- c(G1 = "Gal(b1-3)GalNAc(a1-", G2 = "GlcNAc(b1-4)GlcNAc(a1-")
motifs <- c(M1 = "Gal(b1-", M2 = "GlcNAc(b1-")
have_motifs(glycans, motifs)
#>       M1    M2
#> G1  TRUE FALSE
#> G2 FALSE  TRUE