Check if a Glycan have the Given Motifs

This is a vectorized versions of has_motif(). It checks if a glycan has the given motifs, or all known motifs in the GlycoMotif GlyGen Collection if motifs is missing.

A logical vector is returned. Names of motifs are preserved as names of the output logical vectors. If no names are provided, and the input is a character vector, the names will be the vector itself.

Usage

has_motifs(
  glycan,
  motifs = NULL,
  ...,
  alignments = "substructure",
  ignore_linkages = FALSE
)

Arguments

glycan

A 'glycan_graph' object, or an IUPAC-condensed structure string.

motifs

A list of 'glycan_graph' objects, a character vector of IUPAC-condensed structure strings, or a character vector of known motif names.

...

Ignored.

alignments

A character vector of alignment names for the motifs, either of the same length as motifs, or a single character string. Could be "substructure", "core", "terminal", or "whole". Default to "substructure". If not provided and motifs are known motif names or missing, the alignments in the GlycoMotif GlyGen Collection will be used.

ignore_linkages

A logical value. If TRUE, linkages will be ignored in the comparison.

Value

A logical vector, indicating if the glycan has the motifs.

Examples

library(glyparse)

# Check if a glycan has the given motifs using names.
# Note that the `motifs` vector will become names of the result vector.
glycan <- "Gal(b1-3)GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-"
has_motifs(glycan, c("O-Glycan core 1", "O-Glycan core 2"))
#> O-Glycan core 1 O-Glycan core 2 
#>            TRUE           FALSE 

# Using IUPAC-condensed structure strings.
# Note that names of `motifs` are preserved in the result vector.
motifs <- c(M1 = "Gal(b1-3)GalNAc(a1-", M2 = "GlcNAc(b1-4)GlcNAc(a1-")
has_motifs(glycan, motifs)
#>    M1    M2 
#>  TRUE FALSE 

# Or using 'glycan_graph' objects.
motifs <- list(
  M1 = parse_iupac_condensed("Gal(b1-3)GalNAc(a1-"),
  M2 = parse_iupac_condensed("GlcNAc(b1-4)GlcNAc(a1-")
)
has_motifs(glycan, motifs)
#>    M1    M2 
#>  TRUE FALSE 

# If motifs are missing, use all known motifs in the
# GlycoMotif GlyGen Collection.
has_motifs(glycan)[1:5]
#> Blood group H (type 2) - Lewis y                        i antigen 
#>                            FALSE                            FALSE 
#>                         LacdiNAc                              GT2 
#>                            FALSE                            FALSE 
#> Blood group B (type 1) - Lewis b 
#>                            FALSE