Check if these Glycans have the Given Motif

This is a vectorized version of has_motif(). It checks if a list of glycans (or a character vector of IUPAC-condensed structure strings) have the given motif.

A logical vector is returned. Names of glycans are preserved as names of the output logical vectors. If no names are provided, and the input is a character vector, the names will be the vector itself.

Usage

have_motif(
  glycans,
  motif,
  ...,
  alignment = "substructure",
  ignore_linkages = FALSE
)

Arguments

glycans

A list of 'glycan_graph' objects, or a character vector of IUPAC-condensed structure strings.

motif

A 'glycan_graph' object, an IUPAC-condensed structure string, or a known motif name.

...

Ignored.

alignment

A character string of alignment name for the motif. Could be "substructure", "core", "terminal", or "whole". Default to "substructure". If not provided and motif is a known motif name, the alignment in the GlycoMotif GlyGen Collection will be used.

ignore_linkages

A logical value. If TRUE, linkages will be ignored in the comparison.

Value

A logical vector, indicating if the glycans have the motif.

Examples

glycan <- c(G1 = "Gal(b1-3)GlcNAc", G2 = "Man(b1-4)GlcNAc", G3 = "GlcNAc")
have_motif(glycan, "Gal")
#>    G1    G2    G3 
#>  TRUE FALSE FALSE